+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Discounts with time restrictions may not be invalidated directly in the checkout...
 - provided additional Order-Number is searchable
 - increase line spacing for the error messages
-- Save Payment Data for registered Users 
+- Save Payment Data for registered Users
+- fix inconsistent data for Paylater payment methods 
 
 ## [1.1.3] - 2023-11-14